// lib/api/generated/schemas/subscriptionPlanSliceDtoPlan.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * НеМалика API
 * Маркетплейс компьютерной техники — REST API для веб-витрины (Next.js) и Telegram mini-app. Покупать можно без авторизации; вход через Telegram нужен продавцу, администратору и покупателю, который хочет видеть историю просмотров на всех своих устройствах.
 * OpenAPI spec version: 1.0.0
 */

export type SubscriptionPlanSliceDtoPlan = typeof SubscriptionPlanSliceDtoPlan[keyof typeof SubscriptionPlanSliceDtoPlan];


export const SubscriptionPlanSliceDtoPlan = {
  start: 'start',
  pro: 'pro',
  max: 'max',
} as const;